requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Text processing and embeddings
tiktoken>=0.4.0
langchain>=0.0.200
llama-index>=0.7.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
httpx>=0.24.0

# Optional GPU support (uncomment if using GPU)
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# faiss-gpu>=1.7.4 